Made pivot the new location of entities spawned in the world
The size of the visual area for an entity should only be an indicator to the user about how big it will seem, changing the size should not change where the entity spawned. This is how it works in most other engines I've seen and I believe it's better to have a defined point which can be seen in the editor to represent the translation of the entity, rather than arbitrarily moving them half way up.