fix: use normal sprite for background color instead of tile (#171)
Our solution for background colors was pretty bad due to a previous
limitation with `bevy`'s sprites. We can now use a normal sprite for the
background color instead of a single-tile tilemap. This PR makes that
change.

This also unblocks #169, since it seems the z-indexing issue was
partially due to our weird use background image solution. Though, I
expect bevy_ecs_tilemap to release an `0.11` somewhat soon to make the
3d-iso changes optional.