fix: insert Name component before evaluating `LdtkEntity` (#186)
Currently, if a user provides a custom constructor for `Name` via
`LdtkEntity` registration, it will be overwritten by the plugin since it
automatically inserts a name from the ldtk entity identifier. This
change performs this insertion *before* evaluating `LdtkEntity` so that
a user-provided `Name` will overwrite the plugin-provided `Name`
instead.