ci: upgrade CI to current bevy_github_ci_template and lint more targets/features (#208)
This changes the linting CI drastically to be based off the most recent
[`bevy_github_ci_template`](https://github.com/bevyengine/bevy_github_ci_template/blob/1b5cf1bc15ef33cc76eb2e7ef1d6150504ed8687/.github/workflows/ci.yaml).
This upgrade includes switching to the actively-maintained dtolnay
actions, and implements caching.

This template was updated to have all jobs run against all-targets and
all-features as necessary. This begins to resolve a couple of
code-review pain-points for me, like having to lint the examples/tests
myself.

The template was also updated to have a `check` job. I imagine this was
excluded from the template since checking compilation is implicitly
performed by the test job. However, I think it's a good idea to include
some basic checks that various feature permutations are OK, since
feature selection can greatly impact compilation. For now, this checks
the package with no features (without `--all-targets`, since some of the
examples require features), with default features, and with all
features.

chore: fix clippy issues for all targets (#208)