test: add basic LDtk faker types and use them LevelMetadataAccessor tests (#238)
I would like to have more comprehensive tests, but one obstacle in that
respect has always been that the LDtk json types were very difficult to
construct for tests. They are quite complicated and massive, so writing
custom test data is very time consuming and difficult to maintain. This
seems to me like a good use case for a faker library, so that is what
this PR implements.

You may notice that these faker types are not very sophisticated, and
only edit a few fields. My thoughts are that, as tests need to fake more
fields of these types, they are welcome to improve the fakers. The sorts
of things that are faked here are those that are used in tests for the
new asset types. These fakers are also pretty careful about making the
faked data a little bit realistic.. in particular, all faked levels will
have the "same" set of layers with some per-instance variation much like
a real LDtk file. At the highest level, there are also fakers for
creating an LdtkJson with unloaded levels, and a separate list of loaded
levels, which can be added to an asset store for testing of accessing
external level data in the future.

I wasn't too careful about documenting these new types for now, since
they are all behind `#[cfg(test)]` and do not appear in the library. If
contributors find these types confusing, it may be worth documenting
them in a future PR.