feat: add LdtkJsonWithMetadata type for representing internal- and external-level project data with generics (#242)
I have been working on redesigning the asset types to provide more
convenient access to level data and a more correct memory model of both
internal- and external-level projects. It is useful to represent these
two types of projects with similar-but-different rust types, where we
can implement some methods/traits as exclusive to each case, and other
methods/traits as the same for each case. This is a good use case for
generics. This PR adds a generic `LdtkJsonWithMetadata<L>` type for
this, which stores just enough information to provide convenient access
to loaded level data in either case.

The generic parameter for this type is `L: LevelLocale`, leveraging the
`LevelLocale` traits/types added in a previous PR. This parameter will
be set to `InternalLevels` for internal-levels projects, and
`ExternalLevels` for external-levels projects. The type stores different
level metadata types in either case. These two versions of the type both
implement `RawLevelAccessor` and `LevelMetadataAccessor`, which provides
many convenience methods for accessing raw levels. For accessing "loaded
levels", these types receive different, non-trait implementation
methods. In the internal case, users can use the `loaded_level` methods,
and in the external case, users can use the `external_level` methods.
While these methods are pretty similar to eachother, only the
external-level methods require access to the `Assets<LdtkLevel>` asset
store. This difference in method signature is the main reason that these
two cases are split up like this in the first place.

Note that this type isn't actually used in the plugin yet. In future
PRs, it will be stored within the `LdtkProject` asset, but this change
is pretty major and deserves to be separated.

(Ignore broken doc-links here, they will no longer be broken in future
PRs)