docs: document all-features in docs.rs (#252)
Closes #246 

Now that entire types are being hidden behind features, it's important
that we enable this option so that they are visible in doc.rs. I also
looked into enabling other common features like
`--generate-link-to-definition`, but this seems to be only relevant if
we're publishing the docs ourselves instead of relying on docs.rs, so I
left that out for now.