docs: remove README.md caveat for hot reloading external levels (#253)
Related to #1 

Since bevy's file changewatcher now accepts a `Duration`, running into
the bug caused by level files being temporarily unavailable is either
extremely unlikely or impossible with a reasonable duration. So, I think
it's safe to remove this caveat from the readme saying that users need
to double-save for hot-reloading to work for external levels.