docs: add mdbook with outline and introduction (#261)
Currently, we only have documentation in the form of code examples and
`docs.rs`. While the docs.rs api reference is pretty comprehensive, it
is not a great platform for providing higher-level explanation
documentation, or tutorials, or guides. Some of this does exist in the
docs.rs docs, but it doesn't really belong there and is difficult to
discover. This PR initializes an mdbook in this repository where these
other types of documentation can reside. I've provided an introduction
and an outline. I explain the goals of the book in its introduction. CI
supporting this book will be added in follow-up PRs.