ci: add "Build and test book" job and basic getting started chapter (#263)
In an effort to keep the mdbook up-to-date, we need CI to be able to
tell us if its code snippets don't compile with the current plugin. This
adds such a job to the lint CI, and leverages mdbook-keeper to do so
since we need to be able to depend on 3rd party crates in the book code.

This also adds some basic text to the getting started tutorial in the
book. This text was basically copied from the README.md. This should be
updated in the future to be more tutorial-y, and once it is, we should
update the README.md/lib.rs to link to the book rather than have their
own copy of the getting started guide. This was mostly added in this PR
so that the new CI job has some code to test.