ci: publish book to github pages on new releases (#264)
Now that we have a book as part of our documentation, new builds of this
book should be published for new versions of the plugin, similar to
docs.rs. This PR accomplishes this by adding a workflow to the CI that
publishes the book to github pages on new releases. Furthermore, the
token used by release-please has been updated so that it can trigger
workflows itself (something the default `secrets.GH_TOKEN` is not
capable of).

The workflow has been tested, and a preview of the book is now
available: https://trouv.github.io/bevy_ecs_ldtk