docs: add 0.8 to 0.9 migration guide (#266)
`bevy_ecs_ldtk` 0.9 has many more breaking changes than any other
release so far. I think it would be nice to provide users with a
migration guide, which will document most of the changes they will need
to make in their games to continue using the plugin. The changes I've
chosen to describe here are basically anything that's labeled as a
breaking change in the changelog. I verified that the repo examples
between now and `0.8` haven't had to deal with any other breaking
change, so this should be complete (especially under normal usage).

Note that this does *not* contain a section about updating to 0.12. I
will add that to the 0.12 upgrade PR after this is merged.

Furthermore, the 0.9 doc-tests here are generally marked as `no_run`,
not `ignore`. These should be changed back to `ignore` before releasing
0.9 so that future changes to the plugin don't break the book. It would
be nice to automate this but it's a pretty easy change to do manually as
part of the release process. I think it's still good to keep them as
`no_run` until that time comes, especially for future releases where
migration guides will be built more incrementally.