docs: add tile-based game example w/ a tutorial in the book, replacing getting-started guide (#269)
Closes #51 

The existing getting-started guide wasn't a satisfactory "tutorial"
according to the "grand unified theory of documentation". This PR adds a
more complete-minimal-project tutorial that covers the steps that the
getting-started guide did, as well as steps for setting up the LDtk
project and adding gameplay. This tutorial is associated with a new
example - `tile_based_game`. Most of the code snippets in the tutorial
try to include code sections from the example to make sure these two
match.

I will likely update the `lib.rs` to link to this tutorial instead of
copy-pasting the same getting-started guide in a future PR. I might also
remove the getting-started guide in the `README.md`, but I'm more okay
with it existing there as marketing material.