docs: write *Game Logic Integration* chapter of the book (#279)
This chapter covers the main strategies for integrating game logic with
LDtk projects, and also compares them. These main strategies being
`LdtkEntity`/`LdtkIntCell` registration and the blueprint pattern. In a
future PR, explanations of these concepts in the API reference and
project readme will be removed in favor of linking to this chapter.