ci: publish book to versioned subdirectories (#280)
Closes #274 

This changes the publish book CI to subdirectories of the github pages
branch rather than the root. The directory name is either the tag name
that triggered the publishing, or the branch name (which can only be
`main` due to the workflow triggers). Furthermore, if it is a tag that
triggered the publishing, the `latest` directory is *also* published to.
So the result will be multiple versions of the book published to github
pages, one for each version of the plugin named after that version, plus
two special versions: `latest` and `main`. This will help with
development of the book and allow users of older versions of the plugin
to reference older documentation.