Add diagram of pivot behavior between versions
For now, I'm putting the file in repo with the other image, knowing that it can be moved later.

The image itself needs to be updated for clarity--I tested on my brother, and he was so distracted by the bright shiny object that he totally missed the tiny pivot dot.