build: use a git revision dependency for bevy_ecs_tilemap instead of a git branch patch (#326)
Closes #321

Currently, we're using a patch to depend on `bevy_ecs_tilemap`. This
places a burden on the user to use a similar patch in their own
`Cargo.toml`. We're also using the `main` branch of `bevy_ecs_tilemap`
in this patch. This can lead to issues when updates occur to that branch
of the `bevy_ecs_tilemap` repository, as new users and `cargo update`rs
will pull in any new changes to that repo, even breaking ones. This is
currently affecting our CI.

This PR's changes will ultimately be overwritten soon by an update to
bevy 0.14 (#325), but merging this sooner will help set the standard for
development of this project inbetween `bevy_ecs_tilemap` releases in the
future, and also unblock other PRs whose CI checks are affected (#306).