ci: address new clippy lints in 1.80 (#331)
Rust 1.80 has been released, and new clippy lints are now being checked
in our CI. This PR addresses these lints to uphold the code quality of
this repository and get the clippy CI job passing again.

The new lint that was failing is `doc_lazy_continuation`, so this PR
only really affects docs. This lint was simply `#[allow(...)]`'d in the
`ldtk` module since most of the code in this module is auto-generated
and any manual changes to it need to be repeated whenever it is
regenerated.