docs: refactor platformer example by topic (#329)
The components and systems in the platformer example are confusingly
sorted: components are stored in a separate file from the systems that
use them, and within those files the components & systems are jumbled
chaotically with little rhyme or reason. In a crate example this
complex, this organization scheme makes understanding the code a real
challenge (which is a major reason I couldn't update this example in
#294).

This PR reorganizes the platformer example into modules sorted roughly
by "part of the game" or "game system" (camera system, climbing,
inventory, etc.). Each module contains both the components and ECS
systems specific to that part, to present a complete picture of what the
game system does and how it works.

No game logic has been changed. This PR focuses purely on code quality.