fix: address lifetime lints and use Level lifetimes in LoadedLevel methods (#364)
The first is just new lifetime lints in Rust. These are primarily just
making the lifetime visible. In a couple places, this actually gave me
the opportunity to make the lifetimes slightly more permissive.

The second is that all the LoadedLevel methods would tie the lifetime of
all the references to the LoadedLevel, when really they are only
constrained by the lifetime of Level. This should allow for slightly
more permissive lifetimes when interacting with LoadedLevel.