refactor: upgrade syn to 2.0 (#395)
As per the comment on pr #343, this upgrades syn from 1.0 to 2.0.

Bevy uses syn 2.0 as well, so after this PR is merged, users won't need
to compile both syn 1.0 and 2.0 libraries, which should save a little
bit of compile time.

Tests compile & pass, and examples compile & seem to work the same as
before the upgrade.